fixed issue #39
