+++ plugin-gradle/src/test/kotlin/io/github/baole/konture/plugin/KonturePluginTest.kt
@@ -6,28 +6,11 @@
 
 package io.github.baole.konture.plugin
 
-import io.github.baole.konture.core.DependencyGraphModel
-import io.github.baole.konture.core.KontureConstants
-import io.github.baole.konture.core.LayoutModel
-import org.gradle.api.GradleException
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.internal.project.ProjectInternal
-import org.gradle.api.provider.Provider
-import org.gradle.api.tasks.testing.Test as GradleTestTask
 import org.gradle.testfixtures.ProjectBuilder
-import org.gradle.testkit.runner.GradleRunner
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertNotNull
-import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.io.TempDir
-import java.io.File
-import java.nio.file.Path
-import kotlinx.serialization.json.Json
 
 class KonturePluginTest {
     @Test
@@ -96,164 +79,6 @@ class KonturePluginTest {
     }
 
     @Test
-    fun `external dependency rule detector follows Kotlin test sources`() {
-        val project = ProjectBuilder.builder().build()
-        val source = project.file("src/test/kotlin/ArchitectureTest.kt")
-        source.parentFile.mkdirs()
-        source.writeText("// should().notDependOnExternalLibraries(\"a:b\")\nclass ArchitectureTest")
-        val result = project.layout.buildDirectory.file("konture/requires-dependencies.txt").get().asFile
-        val task = project.tasks.create("detectExternalRules", DetectExternalDependencyRules::class.java)
-        task.testSources.from(source)
-        task.resultFile.set(result)
-
-        task.detect()
-        assertEquals("false", result.readText())
-
-        source.writeText("class ArchitectureTest { fun rule() = should().notDependOnExternalLibraries(\"a:b\") }")
-        task.detect()
-        assertEquals("true", result.readText())
-    }
-
-    @Test
-    fun `root detector includes custom test source sets of Konture consumers`() {
-        val root = ProjectBuilder.builder().withName("root").build()
-        root.plugins.apply("io.github.baole.konture")
-        val consumer = ProjectBuilder.builder().withName("architecture").withParent(root).build()
-        consumer.plugins.apply("io.github.baole.konture")
-        val source = consumer.file("src/commonTest/kotlin/ArchitectureTest.kt")
-        source.parentFile.mkdirs()
-        source.writeText("fun rule() = should().onlyDependOnExternalLibraries(\"a:b\")")
-
-        val detector = root.tasks.getByName("detectKontureExternalDependencyRules") as DetectExternalDependencyRules
-        detector.detect()
-
-        assertEquals("true", detector.resultFile.get().asFile.readText())
-    }
-
-    @Test
-    fun `testGenerateArchitectureLayoutTask`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-
-        // Create child projects
-        val childA =
-            ProjectBuilder
-                .builder()
-                .withName("child-a")
-                .withParent(rootProject)
-                .build()
-        val childB =
-            ProjectBuilder
-                .builder()
-                .withName("child-b")
-                .withParent(rootProject)
-                .build()
-
-        // Apply plugins to child projects
-        childA.plugins.apply("org.jetbrains.kotlin.jvm")
-        childB.plugins.apply("java")
-
-        // Apply konture to root
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        // Create some source files in child-a source dirs to test file walking
-        val childADir = childA.projectDir
-        val srcDir = File(childADir, "src/main/kotlin/com/example")
-        srcDir.mkdirs()
-        val dummyFile = File(srcDir, "Example.kt")
-        dummyFile.writeText("package com.example\nclass Example")
-
-        // Add a project dependency from child-a to child-b
-        childA.configurations.getByName("implementation").dependencies.add(
-            childA.dependencies.project(mapOf("path" to ":child-b")),
-        )
-
-        // Run afterEvaluate hooks to eagerly configure tasks
-        (rootProject as ProjectInternal).evaluate()
-        (childA as ProjectInternal).evaluate()
-        (childB as ProjectInternal).evaluate()
-
-        // Run the task
-        val task = rootProject.tasks.getByName("generateArchitectureLayout") as GenerateArchitectureLayout
-        task.outputFile
-            .get()
-            .asFile.parentFile
-            .mkdirs()
-        task.generate()
-
-        // Check output JSON file
-        val outputFile = task.outputFile.get().asFile
-        assertTrue(outputFile.exists())
-
-        val jsonText = outputFile.readText()
-        val layoutModel =
-            Json.decodeFromString(
-                LayoutModel
-                    .serializer(),
-                jsonText,
-            )
-
-        assertEquals(LayoutModel.CURRENT_SCHEMA_VERSION, layoutModel.schemaVersion)
-        val rootBuild = layoutModel.builds.firstOrNull { it.id == ":" }
-        assertNotNull(rootBuild)
-
-        val moduleA = rootBuild?.modules?.firstOrNull { it.path == ":child-a" }
-        assertNotNull(moduleA)
-        assertTrue(moduleA?.appliedPlugins?.contains("kotlin-jvm") == true)
-
-        val moduleB = rootBuild?.modules?.firstOrNull { it.path == ":child-b" }
-        assertNotNull(moduleB)
-
-        // Verify child-a has Example.kt listed relative to moduleDir
-        val mainSourceSet = moduleA?.sourceSets?.firstOrNull { it.name == "main" }
-        assertNotNull(mainSourceSet)
-        assertTrue(mainSourceSet?.kotlinFiles?.contains("src/main/kotlin/com/example/Example.kt") == true)
-
-        // Verify dependency is collected
-        val dep = moduleA?.dependencies?.firstOrNull { it.targetPath == ":child-b" }
-        assertNotNull(dep)
-        assertEquals("implementation", dep?.configuration)
-    }
-
-    @Test
-    fun `testAbsoluteAndExternalSourceDirs`() {
-        val rootProject = ProjectBuilder.builder().build()
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        val task = rootProject.tasks.getByName("generateArchitectureLayout") as GenerateArchitectureLayout
-
-        // Manually build custom ModuleData with an external source directory
-        val externalDir = File("/some/external/absolute/path") // absolute path not under root
-        val sourceSet =
-            SourceSetData(
-                name = "external",
-                kind = "KOTLIN_JVM",
-                production = true,
-                srcDirs = listOf(externalDir.absolutePath),
-            )
-        val module =
-            ModuleData(
-                path = ":custom",
-                projectDir = rootProject.projectDir.absolutePath,
-                appliedPlugins = listOf("kotlin-jvm"),
-                sourceSets = listOf(sourceSet),
-                dependencies = emptyList(),
-            )
-
-        task.modules.set(listOf(module))
-        task.outputFile
-            .get()
-            .asFile.parentFile
-            .mkdirs()
-        task.generate()
-
-        // Verify the task runs successfully and keeps the absolute path
-        val outputFile = task.outputFile.get().asFile
-        assertTrue(outputFile.exists())
-        val jsonText = outputFile.readText()
-        assertTrue(jsonText.contains("/some/external/absolute/path"))
-    }
-
-    @Test
     fun `testPluginDataClasses`() {
         val sourceSet1 = SourceSetData("main", "KOTLIN_JVM", true, listOf("src"))
         val sourceSet2 = SourceSetData("main", "KOTLIN_JVM", true, listOf("src"))
@@ -300,479 +125,5 @@ class KonturePluginTest {
                 .get()
                 .asFile
         assertEquals(expectedFile, files.first())
-    }
-
-    @Test
-    fun `testExclusionsAreSerializedIntoLayoutJson`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        val extension = rootProject.extensions.getByName("konture") as KontureExtension
-        extension.excludeModules(":module-b")
-        extension.excludePackages("com.example.exclude..")
-        extension.excludeClasses("ExcludedClass")
-
-        val task = rootProject.tasks.getByName("generateArchitectureLayout") as GenerateArchitectureLayout
-        task.outputFile
-            .get()
-            .asFile.parentFile
-            .mkdirs()
-        task.generate()
-
-        val outputFile = task.outputFile.get().asFile
-        assertTrue(outputFile.exists())
-
-        val jsonText = outputFile.readText()
-        val layoutModel =
-            Json.decodeFromString(
-                LayoutModel
-                    .serializer(),
-                jsonText,
-            )
-
-        assertEquals(listOf(":module-b"), layoutModel.exclusions.excludeModules)
-        assertEquals(listOf("com.example.exclude.."), layoutModel.exclusions.excludePackages)
-        assertEquals(listOf("ExcludedClass"), layoutModel.exclusions.excludeClasses)
-    }
-
-    @Test
-    fun `testJavaPluginFallbackSourceSets`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val child =
-            ProjectBuilder
-                .builder()
-                .withName("child-java")
-                .withParent(rootProject)
-                .build()
-        child.plugins.apply("java")
-
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        // Run afterEvaluate hooks to eagerly configure tasks
-        (rootProject as ProjectInternal).evaluate()
-        (child as ProjectInternal).evaluate()
-
-        val task = rootProject.tasks.getByName("generateArchitectureLayout") as GenerateArchitectureLayout
-        task.outputFile
-            .get()
-            .asFile.parentFile
-            .mkdirs()
-        task.generate()
-
-        val outputFile = task.outputFile.get().asFile
-        assertTrue(outputFile.exists())
-
-        val jsonText = outputFile.readText()
-        val layoutModel =
-            Json.decodeFromString(
-                LayoutModel
-                    .serializer(),
-                jsonText,
-            )
-
-        val rootBuild = layoutModel.builds.firstOrNull { it.id == ":" }
-        assertNotNull(rootBuild)
-
-        val moduleJava = rootBuild?.modules?.firstOrNull { it.path == ":child-java" }
-        assertNotNull(moduleJava)
-
-        // Verify that java fallback successfully populated sourceSets
-        val mainSourceSet = moduleJava?.sourceSets?.firstOrNull { it.name == "main" }
-        assertNotNull(mainSourceSet)
-        assertEquals("KOTLIN_JVM", mainSourceSet?.kind?.name)
-        assertTrue(mainSourceSet?.production == true)
-        assertTrue(
-            mainSourceSet?.srcDirs?.any { it.endsWith("src/main/java") || it.endsWith("src/main/resources") } == true,
-        )
-    }
-
-    @Test
-    fun `testKmpSourceSetExtraction`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val kmpProject =
-            ProjectBuilder
-                .builder()
-                .withName("kmp-child")
-                .withParent(rootProject)
-                .build()
-
-        // Apply Kotlin Multiplatform plugin
-        kmpProject.plugins.apply("org.jetbrains.kotlin.multiplatform")
-
-        // Retrieve KMP extension and define targets to register source sets
-        val kotlinExt =
-            kmpProject.extensions.getByType(
-                KotlinMultiplatformExtension::class.java,
-            )
-        kotlinExt.jvm("desktop")
-        kotlinExt.iosX64("ios")
-        // Collect source sets via the helper function inside KonturePlugin
-        val plugin = KonturePlugin()
-        val sourceSets = plugin.collectSourceSets(kmpProject)
-
-        // Verify we found some source sets
-        assertTrue(sourceSets.isNotEmpty())
-
-        // Verify KMP is correctly detected as the kind for all collected source sets
-        for (ss in sourceSets) {
-            assertEquals("KMP", ss.kind)
-        }
-
-        // Verify commonMain is production, and commonTest is NOT production
-        val commonMain = sourceSets.firstOrNull { it.name == "commonMain" }
-        assertNotNull(commonMain)
-        assertTrue(commonMain?.production == true)
-
-        val commonTest = sourceSets.firstOrNull { it.name == "commonTest" }
-        assertNotNull(commonTest)
-        assertTrue(commonTest?.production == false)
-
-        // Verify platform specific targets (desktopMain, iosMain)
-        val desktopMain = sourceSets.firstOrNull { it.name == "desktopMain" }
-        assertNotNull(desktopMain)
-        assertTrue(desktopMain?.production == true)
-
-        val iosTest = sourceSets.firstOrNull { it.name == "iosTest" }
-        assertNotNull(iosTest)
-        assertTrue(iosTest?.production == false)
-    }
-
-    @Test
-    fun `testKmpSourceSetPlatformTargetsWithTypedApi`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val kmpProject =
-            ProjectBuilder
-                .builder()
-                .withName("kmp-child")
-                .withParent(rootProject)
-                .build()
-
-        // Apply Kotlin Multiplatform plugin
-        kmpProject.plugins.apply("org.jetbrains.kotlin.multiplatform")
-
-        // Retrieve KMP extension and define targets to register source sets
-        val kotlinExt =
-            kmpProject.extensions.getByType(
-                KotlinMultiplatformExtension::class.java,
-            )
-        kotlinExt.jvm("desktop")
-        kotlinExt.iosX64("ios")
-        val commonSourceSet = kotlinExt.sourceSets.getByName("commonMain")
-        kotlinExt.sourceSets.getByName("desktopMain").dependsOn(commonSourceSet)
-        kotlinExt.sourceSets.getByName("iosMain").dependsOn(commonSourceSet)
-        val intermediate = kotlinExt.sourceSets.create("intermediateCode")
-        kotlinExt.sourceSets.getByName("desktopMain").dependsOn(intermediate)
-
-        val plugin = KonturePlugin()
-        val sourceSets = plugin.collectSourceSets(kmpProject)
-
-        // Find desktopMain
-        val desktopMain = sourceSets.firstOrNull { it.name == "desktopMain" }
-        assertNotNull(desktopMain)
-        // Verify it extracts platform target metadata correctly!
-        assertTrue(desktopMain?.platforms?.contains("jvm") == true)
-        assertTrue(desktopMain?.dependsOnSourceSets?.contains("commonMain") == true)
-        assertTrue(sourceSets.firstOrNull { it.name == "intermediateCode" }?.production == true)
-
-        val commonMain = sourceSets.firstOrNull { it.name == "commonMain" }
-        assertNotNull(commonMain)
-        assertTrue(commonMain?.targetNames?.isNotEmpty() == true)
-        assertFalse(commonMain?.targetNames?.contains("desktop") == true)
-
-        val iosMain = sourceSets.firstOrNull { it.name == "iosMain" }
-        assertNotNull(iosMain)
-        assertTrue(iosMain?.platforms?.contains("native") == true)
-        assertTrue(iosMain?.targetNames?.isNotEmpty() == true)
-        assertFalse(iosMain?.targetNames?.contains("desktop") == true)
-        assertTrue(iosMain?.dependencyConfigurations?.contains("iosMainImplementation") == true)
-    }
-
-    @Test
-    fun `testSpecializedAndroidPlugins`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val dynamicFeature =
-            ProjectBuilder
-                .builder()
-                .withName("feature")
-                .withParent(rootProject)
-                .build()
-        val testModule =
-            ProjectBuilder
-                .builder()
-                .withName("test-module")
-                .withParent(rootProject)
-                .build()
-
-        class DummyDynamicFeaturePlugin : Plugin<Project> {
-            override fun apply(target: Project) = Unit
-        }
-
-        class DummyAndroidTestPlugin : Plugin<Project> {
-            override fun apply(target: Project) = Unit
-        }
-
-        dynamicFeature.plugins.apply(DummyDynamicFeaturePlugin::class.java)
-        testModule.plugins.apply(DummyAndroidTestPlugin::class.java)
-
-        val plugin = KonturePlugin()
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        val modulesList =
-            rootProject.allprojects.map { sub ->
-                val plugins = mutableListOf<String>()
-                if (sub.pluginManager.hasPlugin("com.android.dynamic-feature") ||
-                    sub.plugins.any { it.javaClass.simpleName.contains("DynamicFeature") }
-                ) {
-                    plugins.add("android-dynamic-feature")
-                }
-                if (sub.pluginManager.hasPlugin("com.android.test") ||
-                    sub.plugins.any {
-                        it.javaClass.simpleName.contains("TestPlugin") ||
-                            sub.plugins.any { it.javaClass.simpleName.contains("AndroidTest") }
-                    }
-                ) {
-                    plugins.add("android-test")
-                }
-                plugins
-            }
-        assertTrue(modulesList.any { it.contains("android-dynamic-feature") })
-        assertTrue(modulesList.any { it.contains("android-test") })
-    }
-
-    @Test
-    fun `testCompositeBuildAndDynamicDependenciesHeuristic`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        rootProject.group = "io.github.baole.konture"
-
-        val child =
-            ProjectBuilder
-                .builder()
-                .withName("child")
-                .withParent(rootProject)
-                .build()
-        child.plugins.apply("org.jetbrains.kotlin.jvm")
-
-        // Add a mock external/composite dependency belonging to the same group prefix
-        child.configurations.getByName("implementation").dependencies.add(
-            rootProject.dependencies.create("io.github.baole.konture:composite-dep:1.0.0"),
-        )
-
-        val plugin = KonturePlugin()
-        val collectDepsMethod =
-            KonturePlugin::class.java.getDeclaredMethod(
-                "collectDependencies",
-                Project::class.java,
-            )
-        collectDepsMethod.isAccessible = true
-        @Suppress("UNCHECKED_CAST")
-        val dependencies = collectDepsMethod.invoke(plugin, child) as List<DependencyData>
-
-        assertTrue(dependencies.any { it.targetPath == ":composite-dep" })
-    }
-
-    @Test
-    fun `testInvalidLogLevelThrowsGradleException`() {
-        val rootProject = ProjectBuilder.builder().build()
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        val task = rootProject.tasks.getByName("generateArchitectureLayout") as GenerateArchitectureLayout
-        task.logLevel.set("INVALID_LOG_LEVEL")
-
-        val exception =
-            assertThrows(GradleException::class.java) {
-                task.generate()
-            }
-        assertTrue(exception.message?.contains("Invalid log level: 'INVALID_LOG_LEVEL'") == true)
-    }
-
-    @Test
-    fun `testGenerateDependencyGraphTask`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val child =
-            ProjectBuilder
-                .builder()
-                .withName("child")
-                .withParent(rootProject)
-                .build()
-        child.plugins.apply("org.jetbrains.kotlin.jvm")
-
-        // Apply our plugin
-        rootProject.plugins.apply("io.github.baole.konture")
-
-        // Evaluate to configure task properties
-        (rootProject as ProjectInternal).evaluate()
-        (child as ProjectInternal).evaluate()
-
-        val task = rootProject.tasks.getByName("generateDependencyGraph") as GenerateDependencyGraph
-        task.outputFile
-            .get()
-            .asFile.parentFile
-            .mkdirs()
-        task.execute()
-
-        val outputFile = task.outputFile.get().asFile
-        assertTrue(outputFile.exists())
-
-        val jsonText = outputFile.readText()
-        val externalDeps =
-            Json.decodeFromString(
-                DependencyGraphModel
-                    .serializer(),
-                jsonText,
-            )
-        assertNotNull(externalDeps)
-    }
-
-    @Test
-    fun `plugin extension configures baseline path`() {
-        val project = ProjectBuilder.builder().build()
-        project.plugins.apply("io.github.baole.konture")
-
-        val extension = project.extensions.getByName("konture") as KontureExtension
-        extension.baselinePath.set("custom-baseline.json")
-
-        assertEquals("custom-baseline.json", extension.baselinePath.get())
-    }
-
-    @Test
-    fun `plugin configures test tasks with baseline path`() {
-        val project = ProjectBuilder.builder().build()
-        project.plugins.apply("java") // Registers Test tasks
-        project.plugins.apply("io.github.baole.konture")
-
-        val extension = project.extensions.getByName("konture") as KontureExtension
-        extension.baselinePath.set("custom-baseline-test.json")
-
-        val testTask = project.tasks.getByName("test") as GradleTestTask
-        val baselinePathProp = testTask.systemProperties["konture.baseline.path"]
-        val resolvedValue =
-            when (baselinePathProp) {
-                is Provider<*> -> baselinePathProp.get()
-                else -> baselinePathProp
-            }
-        assertEquals("custom-baseline-test.json", resolvedValue)
-    }
-
-    @Test
-    fun `test task accepts a missing baseline file`(
-        @TempDir projectDir: Path,
-    ) {
-        projectDir.resolve("settings.gradle.kts").toFile().writeText("rootProject.name = \"missing-baseline\"")
-        projectDir.resolve("build.gradle.kts").toFile().writeText(
-            """
-            plugins {
-                java
-                id("io.github.baole.konture")
-            }
-
-            repositories {
-                mavenCentral()
-            }
-
-            dependencies {
-                testImplementation("org.junit.jupiter:junit-jupiter:5.12.0")
-                testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-            }
-
-            tasks.test {
-                useJUnitPlatform()
-            }
-            """.trimIndent(),
-        )
-        val sourceFile = projectDir.resolve("src/test/java/PassingTest.java").toFile()
-        sourceFile.parentFile.mkdirs()
-        sourceFile.writeText(
-            """
-            import org.junit.jupiter.api.Test;
-
-            class PassingTest {
-                @Test void passes() { }
-            }
-            """.trimIndent(),
-        )
-
-        val result =
-            GradleRunner
-                .create()
-                .withProjectDir(projectDir.toFile())
-                .withPluginClasspath()
-                .withArguments("test", "--stacktrace")
-                .build()
-
-        assertEquals("SUCCESS", result.task(":test")?.outcome?.name)
-    }
-
-    @Test
-    fun `root generateKontureBaseline task aggregates subprojects`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val childProject = ProjectBuilder.builder().withName("child").withParent(rootProject).build()
-
-        rootProject.plugins.apply("io.github.baole.konture")
-        childProject.plugins.apply("io.github.baole.konture")
-
-        val rootTask = rootProject.tasks.getByName("generateKontureBaseline")
-        val childTask = childProject.tasks.getByName("generateKontureBaseline")
-
-        val resolvedDeps = rootTask.taskDependencies.getDependencies(rootTask)
-        assertTrue(resolvedDeps.contains(childTask))
-    }
-
-    @Test
-    fun `cli system property override wins over Gradle DSL extension`() {
-        val project = ProjectBuilder.builder().build()
-        project.plugins.apply("java") // Registers Test tasks
-
-        System.setProperty(KontureConstants.PROPERTY_BASELINE_PATH, "cli-override-baseline.json")
-        System.setProperty(KontureConstants.PROPERTY_BASELINE_DIR, "/cli-override-dir")
-        try {
-            project.plugins.apply("io.github.baole.konture")
-
-            val extension = project.extensions.getByName("konture") as KontureExtension
-            extension.baselinePath.set("dsl-baseline.json")
-
-            val testTask = project.tasks.getByName("test") as GradleTestTask
-
-            // Check baseline path
-            val baselinePathProp = testTask.systemProperties[KontureConstants.PROPERTY_BASELINE_PATH]
-            val resolvedPath =
-                when (baselinePathProp) {
-                    is Provider<*> -> baselinePathProp.get()
-                    else -> baselinePathProp
-                }
-            assertEquals("cli-override-baseline.json", resolvedPath)
-
-            // Check baseline dir
-            val baselineDirProp = testTask.systemProperties[KontureConstants.PROPERTY_BASELINE_DIR]
-            val resolvedDir =
-                when (baselineDirProp) {
-                    is Provider<*> -> baselineDirProp.get()
-                    else -> baselineDirProp
-                }
-            assertEquals("/cli-override-dir", resolvedDir)
-        } finally {
-            System.clearProperty(KontureConstants.PROPERTY_BASELINE_PATH)
-            System.clearProperty(KontureConstants.PROPERTY_BASELINE_DIR)
-        }
-    }
-
-    @Test
-    fun `qualified root invocation enables generate mode in subproject test task`() {
-        val rootProject = ProjectBuilder.builder().withName("root").build()
-        val childProject = ProjectBuilder.builder().withName("child").withParent(rootProject).build()
-
-        rootProject.gradle.startParameter.setTaskNames(listOf(":generateKontureBaseline"))
-
-        rootProject.plugins.apply("io.github.baole.konture")
-        childProject.plugins.apply("java") // Registers Test tasks
-        childProject.plugins.apply("io.github.baole.konture")
-
-        val childTestTask = childProject.tasks.getByName("test") as GradleTestTask
-        val generateProp = childTestTask.systemProperties[KontureConstants.PROPERTY_BASELINE_GENERATE]
-        val resolvedValue =
-            when (generateProp) {
-                is Provider<*> -> generateProp.get()
-                else -> generateProp
-            }
-        assertEquals("true", resolvedValue.toString())
     }
 }
